refactor: Add parser.ts (exports parsePSVG function)
